Removed original audio to spectro.
